YARN-11511 Improve TestRMWebServices test config and data
move some class to util
